Add ability to auto-detect line endings.
When a CSV is created on a Mac, the line endings aren't always used by
fgetcsv. We account for this by setting auto_detect_line_endings to true
in the PHP ini.

By using ini_set() this setting is updated for this script's run and
then set back to the standard PHP ini on the server.